Finished link formatting
